change menu.java,MenuRepository.java and add Menu Controller(/api/menu)

+++ src/main/java/com/example/demo/repo/MenuRepository.java
@@ -1,9 +1,6 @@
 package com.example.demo.repo;
 
-import com.example.demo.model.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import com.example.demo.model.Menu;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer>{
-
-}
+public interface MenuRepository extends JpaRepository<Menu, Long> { }
